refactor: extract PRCoordinator from RunwayStore (P1)
Move all PR state, polling, enrichment, actions, and detail caching
into a separate @Observable PRCoordinator class. This reduces
RunwayStore from 1818 → 1391 LOC and isolates PR property mutations
so they no longer invalidate terminal, sidebar, and session views.

PRCoordinator (496 LOC) owns:
- 12 @Observable state properties (pullRequests, selectedPRID, prDetail,
  prTab, sessionPRs, sessionPRIDs, reviewPRCandidate, etc.)
- All PR methods: fetch, enrich, poll, link, select, approve, merge,
  comment, requestChanges, close, toggleDraft, updateBranch, autoMerge
- Polling tasks (30s PR poll + 15s session-PR link)
- Detail cache with 5-minute TTL

RunwayStore keeps a `let prCoordinator: PRCoordinator` reference.
PRCoordinator uses a weak back-reference for cross-cutting concerns
(statusMessage, currentView navigation, session access).

RunwayApp updated: all `store.prXxx` references → `store.prCoordinator.prXxx`.
SidebarActions conformance delegates selectPR/reviewPR to prCoordinator.
